# frontend/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LineSight",
    "description": "LineSight: AI-driven Semantic ETL platform for SMB apparel manufacturing. Parse messy Excel files, track SAM/DHU metrics, and ensure UFLPA compliance.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/dev/update-limits": {
      "put": {
        "tags": [
          "Development"
        ],
        "summary": "Update Limits",
        "description": "Update organization limits for a specific user (email).\nUseful for testing quota enforcement.",
        "operationId": "update_limits_api_v1_dev_update_limits_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LimitUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dev/aggregate-dhu": {
      "post": {
        "tags": [
          "Development"
        ],
        "summary": "Trigger Dhu Aggregation",
        "description": "Manually trigger DHU report aggregation.\n\nComputes DHUReport entries from QualityInspection data.\nUseful for testing and backfilling.\n\nArgs:\n    days_back: Number of days to process (default 7)\n    factory_id: Optional filter for specific factory",
        "operationId": "trigger_dhu_aggregation_api_v1_dev_aggregate_dhu_post",
        "parameters": [
          {
            "name": "days_back",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 7,
              "title": "Days Back"
            }
          },
          {
            "name": "factory_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Factory Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LimitUpdate": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email",
            "default": "demo@linesight.io"
          },
          "max_factories": {
            "type": "integer",
            "title": "Max Factories"
          },
          "max_lines_per_factory": {
            "type": "integer",
            "title": "Max Lines Per Factory"
          },
          "subscription_tier": {
            "$ref": "#/components/schemas/SubscriptionTier",
            "default": "starter"
          }
        },
        "type": "object",
        "required": [
          "max_factories",
          "max_lines_per_factory"
        ],
        "title": "LimitUpdate"
      },
...
      "SubscriptionTier": {
        "type": "string",
        "enum": [
          "starter",
          "pro",
          "enterprise"
        ],
        "title": "SubscriptionTier",
        "description": "Organization subscription tiers."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}